Moving EntityScreen from ‘components’ to ‘screens’
