+++ ui/src/app/Router.jsx
@@ -8,8 +8,8 @@ import LoginScreen from 'src/components/auth/LoginScreen';
 import LogoutScreen from 'src/components/auth/LogoutScreen';
 import SignupScreen from 'src/components/auth/SignupScreen';
 import ActivateScreen from 'src/components/auth/ActivateScreen';
-import EntityScreen from 'src/components/EntityScreen';
-import EntityRelatedScreen from 'src/components/EntityScreen/EntityRelatedScreen';
+import EntityScreen from 'src/screens/EntityScreen';
+import EntityRelatedScreen from 'src/screens/EntityScreen/EntityRelatedScreen';
 import DocumentScreen from 'src/components/DocumentScreen';
 import DocumentRelatedScreen from 'src/components/DocumentScreen/DocumentRelatedScreen';
 import DocumentRedirectScreen from 'src/components/DocumentScreen/DocumentRedirectScreen';
