+++ ui/src/components/DocumentScreen/index.jsx
@@ -5,7 +5,7 @@ import { defineMessages, injectIntl } from 'react-intl';
 
 import { fetchDocument } from 'src/actions';
 import Screen from 'src/components/common/Screen';
-import Entity from 'src/components/EntityScreen/Entity';
+import Entity from 'src/screens/EntityScreen/Entity';
 import Breadcrumbs from 'src/components/common/Breadcrumbs';
 import ScreenLoading from 'src/components/common/ScreenLoading';
 import DualPane from 'src/components/common/DualPane';
