+++ ui/src/components/EntityTable/EntityTableRow.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import Country from 'src/components/common/Country';
 import Schema from 'src/components/common/Schema';
 import Collection from 'src/components/common/Collection';
-import Entity from 'src/components/EntityScreen/Entity';
+import Entity from 'src/screens/EntityScreen/Entity';
 import FileSize from 'src/components/common/FileSize';
 import Date from 'src/components/common/Date';
