+++ ui/src/screens/EntityScreen/Entity.jsx
