+++ ui/src/screens/EntityScreen/Entity.scss
