+++ ui/src/screens/EntityScreen/EntityContent.jsx
