+++ ui/src/screens/EntityScreen/EntityContent.scss
