+++ ui/src/screens/EntityScreen/EntityReferencesTable.scss
