+++ ui/src/screens/EntityScreen/Property.jsx
