+++ ui/src/screens/EntityScreen/Property.scss
